requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
aiosqlite==0.19.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0

# Redis
redis==5.0.1

# Audio Processing
faster-whisper==1.0.0
pyannote.audio==3.1.1
pydub==0.25.1
librosa==0.10.1

# LLM Clients
openai==1.12.0
httpx==0.26.0

# Utils
python-ulid==1.1.0
tenacity==8.2.3
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3